se arregla varios errores de mails y products

+++ src/controller/products.controller.js
@@ -1,10 +1,11 @@
-import productService from "../service/products.service.js";
-import UserController from "../controller/users.controller.js";
-import { JWT_SECRET, verificarToken } from "../utils.js";
+import productService from '../service/products.service.js';
+import UserController from '../controller/users.controller.js';
+import { JWT_SECRET, verificarToken } from '../utils.js';
+import emailService from '../service/email.service.js';
 
 const buildResponse = (data) => {
   return {
-    status: "success",
+    status: 'success',
     payload: data.docs.map((product) => product.toJSON()),
     totalPages: data.totalPages,
     prevPage: data.prevPage,
@@ -15,26 +16,26 @@ const buildResponse = (data) => {
     prevLink: data.hasPrevPage
       ? `http://localhost:8080/api/products?limit=${data.limit}&page=${
           data.prevPage
-        }${data.category ? `&category=${data.category}` : ""}${
-          data.sort ? `&sort=${data.sort}` : ""
+        }${data.category ? `&category=${data.category}` : ''}${
+          data.sort ? `&sort=${data.sort}` : ''
         }`
-      : "",
+      : '',
     nextLink: data.hasNextPage
       ? `http://localhost:8080/api/products?limit=${data.limit}&page=${
           data.nextPage
-        }${data.category ? `&category=${data.category}` : ""}${
-          data.sort ? `&sort=${data.sort}` : ""
+        }${data.category ? `&category=${data.category}` : ''}${
+          data.sort ? `&sort=${data.sort}` : ''
         }`
-      : "",
+      : '',
   };
 };
 
 export default class ProductController {
-  async get(req, res) {
+  static async get(req, res) {
     const { page = 1, limit = 10, category, sort } = req.query;
     const opts = { page, limit };
     if (sort) {
-      opts.sort = { price: sort === "asc" ? 1 : sort === "desc" ? -1 : 1 };
+      opts.sort = { price: sort === 'asc' ? 1 : sort === 'desc' ? -1 : 1 };
     }
     const criteria = {};
     if (category) {
@@ -51,25 +52,74 @@ export default class ProductController {
       const dataForRendering = { ...build, user, cartId };
       return dataForRendering;
     } catch (err) {
-      console.log(err, "error");
+      console.log(err, 'error');
     }
   }
 
-  async create(body, req, res) {
-    const product = await productService.create(body);
-    return product;
+ static async create(body) {
+    try {
+      const product = await productService.create(body);
+      return product;
+    } catch (error) {
+      console.error('Error en el método create:', error);
+      throw new Error('Error interno del servidor');
+    }
   }
 
-  async getById(pid) {
-    const product = await productService.getById(pid);
-    return product;
+  static async getById(pid) {
+    try {
+      const product = await productService.getById(pid);
+      if (!product) {
+        throw new Error('Producto no encontrado');
+      }
+      return product;
+    } catch (error) {
+      console.error('Error en el método getById:', error);
+      throw new Error('Error interno del servidor');
+    }
   }
 
-  async updateById(req, res, pid, body) {
-    await productService.updateById(pid, body);
+ static async updateById(pid, body) {
+    try {
+      await productService.updateById(pid, body);
+      return 'Producto actualizado exitosamente';
+    } catch (error) {
+      console.error('Error en el método updateById:', error);
+      throw new Error('Error interno del servidor');
+    }
   }
 
-  async deleteById(req, res, pid) {
-    await productService.deleteById(pid);
+ static async deleteById(pid, req) {
+    try {
+      const userSession = req.session.user;
+      const decodedToken = verificarToken(userSession, JWT_SECRET);
+      const newuser = await UserController.getById(decodedToken.id);
+      const product = await productService.getById(pid);
+      if (!product) {
+        throw new Error('Producto no encontrado');
+      }
+      //SI el producto no tiene dueño o es ADMIN o USER se borra directamente
+      if (
+        !product.owner ||
+        newuser.role === 'ADMIN' ||
+        newuser.role === 'USER'
+      ) {
+        await productService.deleteById(pid);
+      }
+      const user = await UserController.getByEmail(product.owner);
+      //si el usuario tiene rol premium se le va enviar un mail con info de lo borrado
+      if (user.role === 'PREMIUM') {
+        await emailService.sendAlertDeleteProduct(user);
+      }
+      //si el usuario es admin o es el mismo dueño del producto puede borrar el producto
+      if (newuser._id === product.owner || newuser.role === 'ADMIN') {
+        await productService.deleteById(pid);
+      }
+
+      return 'Producto eliminado exitosamente';
+    } catch (error) {
+      console.error('Error en el método deleteById:', error);
+      throw new Error('Error interno del servidor');
+    }
   }
 }
